requirements.txt
streamlit==1.28.1
langchain==0.1.0
chromadb==0.4.18
openai-whisper==20231117
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
ollama==0.1.7